make hp double, gain 1% hp per move

+++ dat.h
@@ -105,7 +105,7 @@ struct Monster
 	long turns;
 
 	/* current hp */
-	long hp;
+	double hp;
 
 	/* armor class */
 	long ac;
